--- src/modules/Elsa.Workflows.Runtime/Instructions/TriggerWorkflowInstruction.cs
@@ -1,1 +1,0 @@
-namespace Elsa.Workflows.Runtime.Instructions;
