+++ src/modules/Elsa.Workflows.Runtime/LogPersistence/Contracts/IActivityPropertyLogPersistenceEvaluator.cs
@@ -1,0 +1,8 @@
+namespace Elsa.Workflows.Runtime;
+
+public interface IActivityPropertyLogPersistenceEvaluator
+{
+    Task<ActivityLogPersistenceModeMap> EvaluateLogPersistenceModesAsync(ActivityExecutionContext context);
+    
+    Task<Dictionary<string, object?>> GetPersistableOutputAsync(ActivityExecutionContext context);
+}
