+++ src/modules/Elsa.Workflows.Runtime/LogPersistence/Extensions/ActivityExecutionPipelineBuilderExtensions.cs
@@ -1,5 +1,6 @@
 using Elsa.Workflows;
 using Elsa.Workflows.Pipelines.ActivityExecution;
+using Elsa.Workflows.Runtime.Middleware;
 using Elsa.Workflows.Runtime.Middleware.Activities;
 
 // ReSharper disable once CheckNamespace
