+++ src/modules/Elsa.Workflows.Runtime/LogPersistence/Models/ActivityLogPersistenceModeMap.cs
@@ -1,0 +1,9 @@
+using Elsa.Workflows.LogPersistence;
+
+namespace Elsa.Workflows.Runtime;
+
+public class ActivityLogPersistenceModeMap
+{
+    public IDictionary<string, LogPersistenceMode> Inputs { get; set; } = new Dictionary<string, LogPersistenceMode>();
+    public IDictionary<string, LogPersistenceMode> Outputs { get; set; } = new Dictionary<string, LogPersistenceMode>();
+}
