+++ src/modules/Elsa.Workflows.Runtime/LogPersistence/Models/LogPersistenceEvaluationMode.cs
